+++ src/types/cartItem.ts
@@ -1,0 +1,13 @@
+export interface Product {
+  id: number;
+  name: string;
+  price: number;
+  imageUrl: string;
+  category: string;
+}
+
+export interface CartItem {
+  id: number;
+  product: Product;
+  quantity: number;
+}
